requirements.txt
Django==2.1.3
django-thumbs-v2==0.4.1
django-widget-tweaks==1.4.3
djangorestframework==3.9.0
Pillow==5.3.0
pytz==2018.7
